# terraform/openapi.yaml
...
swagger: "2.0"
info:
  title: actividad-api
  version: "1.0"
schemes:
  - https
...
paths:
...
  /users/{id}:
    parameters:
      - in: path
        name: id
        required: true
        type: string
    get:
      operationId: verUsuario
      responses:
        "200":
          description: Un usuario
    put:
      operationId: editarUsuario
      responses:
        "200":
          description: Usuario editado
    delete:
      operationId: borrarUsuario
      responses:
        "200":
          description: Usuario borrado
...
  /upload/{id}:
    parameters:
      - in: path
        name: id
        required: true
        type: string
    delete:
      operationId: borrarArchivo
      responses:
        "200":
          description: Archivo borrado
...